\name{tensorMCCA-package}
\alias{tensorMCCA-package}
\alias{tensorMCCA}
\docType{package}
\title{
\packageTitle{tensorMCCA}
}

\description{
\packageDescription{tensorMCCA}
}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{tensorMCCA}
\packageIndices{tensorMCCA}
}

\author{
\packageAuthor{tensorMCCA}

Maintainer: \packageMaintainer{tensorMCCA}
}


% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }

\examples{
~~ simple examples of the most important functions ~~
}
